refactor plus fcr create
